Move ValueDriver to analysis, clean up ast driver stuff
