+++ source/ast/CMakeLists.txt
@@ -60,6 +60,7 @@ target_sources(
           FmtHelpers.cpp
           HierarchicalReference.cpp
           Lookup.cpp
+          LSPUtilities.cpp
           LValue.cpp
           OpaqueInstancePath.cpp
           Patterns.cpp
