+++ source/ast/expressions/MiscExpressions.cpp
@@ -266,22 +266,6 @@ bool ValueExpressionBase::requireLValueImpl(const ASTContext& context, SourceLoc
         }
 
         if (auto expr = modportPort.getConnectionExpr()) {
-            // We want to record this redirection with the compilation as a side
-            // effect of the instance port used to get here. Normally that gets
-            // done in the ValueSymbol::addDriver call but because we redirect
-            // here we'll lose that information so need to do it manually.
-            if (!context.flags.has(ASTFlags::NotADriver) && !context.scope->isUninstantiated() &&
-                kind == ExpressionKind::HierarchicalValue && !longestStaticPrefix) {
-
-                auto& hve = as<HierarchicalValueExpression>();
-                if (hve.ref.isViaIfacePort()) {
-                    auto& comp = context.getCompilation();
-                    auto driver = comp.emplace<ValueDriver>(context.getDriverKind(), *this,
-                                                            context.getContainingSymbol(), flags);
-                    comp.noteInterfacePortDriver(hve.ref, *driver);
-                }
-            }
-
             // The assignment is actually to the underlying connection expression,
             // so redirect it there.
             return expr->requireLValue(context, location, flags, longestStaticPrefix);
@@ -787,9 +771,6 @@ bool AssertionInstanceExpression::checkAssertionArg(const PropertyExprSyntax& pr
             ctx.addDiag(diag::AssertionOutputLocalVar, bound.sourceRange);
             return false;
         }
-
-        sym->as<ValueSymbol>().addDriver(DriverKind::Procedural, bound, context.scope->asSymbol(),
-                                         AssignFlags::AssertionLocalVarFormalArg);
     }
 
     result = &bound;
